CPR-305 move CourtCaseEventListenerIntegrationTest to use multi node base test
