+++ src/test/kotlin/uk/gov/justice/digital/hmpps/personrecord/integration/MultiNodeTestBase.kt
@@ -3,7 +3,9 @@ package uk.gov.justice.digital.hmpps.personrecord.integration
 import com.microsoft.applicationinsights.TelemetryClient
 import org.assertj.core.api.Assertions.assertThat
 import org.awaitility.kotlin.await
+import org.awaitility.kotlin.matches
 import org.awaitility.kotlin.untilAsserted
+import org.awaitility.kotlin.untilCallTo
 import org.junit.jupiter.api.AfterAll
 import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.BeforeEach
@@ -12,11 +14,17 @@ import org.mockito.kotlin.times
 import org.mockito.kotlin.verify
 import org.mockito.verification.VerificationMode
 import org.springframework.boot.builder.SpringApplicationBuilder
+import software.amazon.awssdk.services.sns.model.MessageAttributeValue
+import software.amazon.awssdk.services.sns.model.PublishRequest
+import software.amazon.awssdk.services.sns.model.PublishResponse
 import software.amazon.awssdk.services.sqs.model.PurgeQueueRequest
 import uk.gov.justice.digital.hmpps.personrecord.HmppsPersonRecord
 import uk.gov.justice.digital.hmpps.personrecord.jpa.repository.PersonRepository
+import uk.gov.justice.digital.hmpps.personrecord.model.hmcts.MessageType
 import uk.gov.justice.digital.hmpps.personrecord.service.type.TelemetryEventType
+import uk.gov.justice.hmpps.sqs.HmppsQueue
 import uk.gov.justice.hmpps.sqs.HmppsQueueService
+import uk.gov.justice.hmpps.sqs.countMessagesOnQueue
 import java.time.Duration
 
 abstract class MultiNodeTestBase {
@@ -53,6 +61,32 @@ abstract class MultiNodeTestBase {
         eq(null),
       )
     }
+  }
+
+  internal fun publishHMCTSMessage(message: String, messageType: MessageType): String {
+    val publishRequest = PublishRequest.builder()
+      .topicArn(courtCaseEventsTopic?.arn)
+      .message(message)
+      .messageAttributes(
+        mapOf(
+          "messageType" to MessageAttributeValue.builder().dataType("String")
+            .stringValue(messageType.name).build(),
+          "messageId" to MessageAttributeValue.builder().dataType("String")
+            .stringValue("d3242a9f-c1cd-4d16-bd46-b7d33ccc9849").build(),
+        ),
+      )
+      .build()
+
+    val response: PublishResponse? = courtCaseEventsTopic?.snsClient?.publish(publishRequest)?.get()
+
+    expectNoMessagesOn(courtCaseEventsQueue)
+    return response!!.messageId()
+  }
+
+  private fun expectNoMessagesOn(queue: HmppsQueue?) {
+    await untilCallTo {
+      queue?.sqsClient?.countMessagesOnQueue(queue.queueUrl)?.get()
+    } matches { it == 0 }
   }
 
   @BeforeEach
